requirement.txt
torch==2.7.1
fair-esm==2.0.0
numpy==2.0.2
pandas==2.3.0
matplotlib==3.9.4
biopython==1.85
timm==1.0.16
ptflops==0.7.4
scikit-learn==1.6.1
tqdm==4.67.1